Add semester viewModel
